refactor: centralize frontend stream helpers
